alignment with modular master
This update changes the project structure towards easier updates. Particularly, it integrates components and api.js in line with progenetix-web, and moves most site specific code & configs to `site-specific`. Now, besides CSS ... the content specific for a project should be in `site-specific` and in `pages`.

These changes are probably incomplete & can use refinements so more re-shuffling to be expected.